+++ libs/actions.close-drawer/src/lib/types.ts
@@ -1,10 +1,4 @@
 import { ActionConfig } from '@spryker/actions';
 
-declare module '@spryker/actions' {
-  interface ActionsRegistry {
-    close_drawer: CloseDrawerActionConfig;
-  }
-}
-
 // tslint:disable-next-line: no-empty-interface
 export interface CloseDrawerActionConfig extends ActionConfig {}