+++ libs/actions.redirect/src/lib/types.ts
@@ -1,11 +1,5 @@
 import { ActionConfig } from '@spryker/actions';
 
-declare module '@spryker/actions' {
-  interface ActionsRegistry {
-    redirect: RedirectActionConfig;
-  }
-}
-
 export interface RedirectActionConfig extends ActionConfig {
   url: string;
 }