+++ libs/actions.refresh-parent-table/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.1](https://github.com/spryker/ui-components/compare/@spryker/actions.refresh-parent-table@0.1.0-next.0...@spryker/actions.refresh-parent-table@0.1.0-next.1) (2021-07-08)
+
+
+### Bug Fixes
+
+* **actions.refresh-parent-table:** fix action name ([8f391e0](https://github.com/spryker/ui-components/commit/8f391e0cff88b7ae138944a16b1a32ffd001af9f))
+
+
+
+
+
 # 0.1.0-next.0 (2021-07-07)