+++ libs/actions.refresh-table/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-next.1](https://github.com/spryker/ui-components/compare/@spryker/actions.refresh-table@0.1.0-next.0...@spryker/actions.refresh-table@0.1.0-next.1) (2021-07-08)
+
+
+### Bug Fixes
+
+* **actions.refresh-table:** fix action name ([5ed8d75](https://github.com/spryker/ui-components/commit/5ed8d750c2855bb427340465519c51c7ec7c6651))
+
+
+
+
+
 # 0.1.0-next.0 (2021-07-07)